+++ messages/deploy.async.md
@@ -8,12 +8,12 @@ Deploy has been queued for cancellation.
 
 # info.AsyncDeployCancel
 
-Run "sf deploy metadata cancel --job-id %s" to cancel the deploy.
+Run "sf project deploy cancel --job-id %s" to cancel the deploy.
 
 # info.AsyncDeployStatus
 
-Run "sf deploy metadata report --job-id %s" to get the latest status.
+Run "sf project deploy report --job-id %s" to get the latest status.
 
 # info.AsyncDeployResume
 
-Run "sf deploy metadata resume --job-id %s" to resume watching the deploy.
+Run "sf project deploy resume --job-id %s" to resume watching the deploy.
