+++ messages/deploy.metadata.md
@@ -98,7 +98,7 @@ Number of minutes to wait for command to complete and display results.
 
 # flags.wait.description
 
-If the command continues to run after the wait period, the CLI returns control of the terminal window to you and returns the job ID. To resume the deployment, run "sf deploy metadata resume". To check the status of the deployment, run "sf deploy metadata report".
+If the command continues to run after the wait period, the CLI returns control of the terminal window to you and returns the job ID. To resume the deployment, run "sf project deploy resume". To check the status of the deployment, run "sf project deploy report".
 
 # flags.manifest.summary
 
@@ -166,7 +166,7 @@ Run the command asynchronously.
 
 # flags.async.description
 
-The command immediately returns the job ID and control of the terminal to you. This way, you can continue to use the CLI. To resume the deployment, run "sf deploy metadata resume". To check the status of the deployment, run "sf deploy metadata report".
+The command immediately returns the job ID and control of the terminal to you. This way, you can continue to use the CLI. To resume the deployment, run "sf project deploy resume". To check the status of the deployment, run "sf project deploy report".
 
 # flags.metadata-dir.summary
 
@@ -191,7 +191,7 @@ You must specify tests using the --tests flag if the --test-level flag is set to
 
 # error.ClientTimeout
 
-The command has timed out, although the deployment is still running. Use "sf deploy metadata resume" to resume watching the deployment.
+The command has timed out, although the deployment is still running. Use "sf project deploy resume" to resume watching the deployment.
 
 # error.Conflicts
 
@@ -201,7 +201,7 @@ There are changes in the org that conflict with the local changes you're trying 
 
 - To overwrite the remote changes, rerun this command with the --ignore-conflicts flag.
 
-- To overwrite the local changes, run the "sf retrieve metadata" command with the --ignore-conflicts flag.
+- To overwrite the local changes, run the "sf project retrieve start" command with the --ignore-conflicts flag.
 
 # error.nothingToDeploy
 
@@ -209,4 +209,4 @@ No local changes to deploy.
 
 # error.nothingToDeploy.Actions
 
-- To see conflicts and ignored files, run "sf deploy metadata preview" with any of the manifest, directory, or metadata flags.
+- To see conflicts and ignored files, run "sf project deploy preview" with any of the manifest, directory, or metadata flags.
