+++ src/commands/project/deploy/cancel.ts
@@ -22,6 +22,8 @@ export default class DeployMetadataCancel extends SfCommand<DeployResultJson> {
   public static readonly examples = messages.getMessages('examples');
   public static readonly requiresProject = true;
   public static readonly state = 'beta';
+  public static readonly aliases = ['deploy:metadata:cancel'];
+  public static readonly deprecateAliases = true;
 
   public static readonly flags = {
     async: Flags.boolean({
@@ -42,6 +44,8 @@ export default class DeployMetadataCancel extends SfCommand<DeployResultJson> {
       summary: messages.getMessage('flags.use-most-recent.summary'),
       exactlyOne: ['use-most-recent', 'job-id'],
     }),
+    // we want to allow undefined to use the value from the cache
+    // eslint-disable-next-line sf-plugin/flag-min-max-default
     wait: Flags.duration({
       char: 'w',
       summary: messages.getMessage('flags.wait.summary'),
@@ -54,8 +58,7 @@ export default class DeployMetadataCancel extends SfCommand<DeployResultJson> {
   };
 
   public async run(): Promise<DeployResultJson> {
-    const { flags } = await this.parse(DeployMetadataCancel);
-    const cache = await DeployCache.create();
+    const [{ flags }, cache] = await Promise.all([this.parse(DeployMetadataCancel), DeployCache.create()]);
     const jobId = cache.resolveLatest(flags['use-most-recent'], flags['job-id']);
 
     const deployOpts = cache.get(jobId);
